# Conflicts:
#	src/main/purescript/Functions/State.purs

+++ src/main/scala/is/clipperz/backend/functions/FileSystem.scala
@@ -5,7 +5,5 @@ import java.nio.file.Files
 
 object FileSystem:
   def deleteAllFiles(file: File): Unit =
-  if file.isDirectory() then
-    file.listFiles.nn.map(_.nn).foreach(deleteAllFiles(_))
-  else
-    Files.deleteIfExists(file.toPath())
+    if file.isDirectory() then file.listFiles.nn.map(_.nn).foreach(deleteAllFiles(_))
+    else Files.deleteIfExists(file.toPath())
